requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
disnake==2.2.2
idna==3.3
mojang==0.2.0
MojangAPI==0.0.5
multidict==5.2.0
python-decouple==3.5
requests==2.26.0
soupsieve==2.3.1
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2